# engines/contracts/engine-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Curf Engine API
  version: 1.0.0
  summary: Data plane for Curf - connections, governed views, queries, policy and audit.
  description: |
    Contract for the Curf engine. Every implementation must pass the shared conformance suite in
    `engines/conformance`. Errors are RFC 9457 `application/problem+json` with a stable `code`.
    Milestone M0 covers identity, audit and operations; later milestones add connections, views,
    queries, reports, exports, sharing and schedules.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Curf engine maintainers
    url: https://github.com/EnterpriseX-Platform/CurfAI
servers:
  - url: http://localhost:8080/engine/v1
    description: Local engine
...
security:
  - bearerAuth: []
paths:
...
  /reports/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getReport
      tags: [Reports]
      summary: Get a report
      description: |
        Editors get the stored definition. Someone who may only run the report gets it for rendering, with each query
        reduced to its `id` and `name`, so SQL, view bindings and data source ids are not handed out. Reports a person may
        not run answer 404.
      responses:
        "200":
          description: The report.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Report" }
                  - { $ref: "#/components/schemas/RunnableReport" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateReport
      tags: [Reports]
      summary: Save a new version of a report
      description: Requires `report:edit`. `version` must be the current one, otherwise 409. Every save keeps a numbered copy.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportRequest" }
      responses:
        "200":
          description: Saved as the next version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/InvalidInput" }
    delete:
      operationId: deleteReport
      tags: [Reports]
      summary: Delete a report
      description: Requires `report:edit`. Its versions are deleted with it.
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /reports/{id}/run:
    post:
      operationId: runReport
      tags: [Reports]
      summary: Run a report as the caller
      description: |
        Runs every query as the caller, so a viewer's row rules and masking apply. Follows Curf's runner: every declared
        parameter is bound (its default, or an empty string), and queries used only by a drill-down are skipped. One query
        failing does not fail the run: its dataset is empty and its provenance carries `executionError` (or
        `accessDeniedNote` when the caller may not use that data source). The answer is in the shape of Curf's own
        run response.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                params:
                  type: object
                  description: Values by parameter name. Unknown names, malformed values and missing required values are 422.
                  additionalProperties: { type: [string, number, boolean, "null"] }
      responses:
        "200":
          description: The datasets and where they came from.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunResult" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/InvalidInput" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthenticated:
      description: Missing, invalid or untrusted token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The caller lacks the required permission.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidInput:
      description: The request is malformed or fails validation. For SQL the code is `CURF_SQL_REJECTED` and `errors` says why.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such resource in the caller's tenant.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: The resource changed since it was read (`CURF_STALE_VERSION`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, examples: ["urn:curf:error:forbidden"] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            - CURF_INVALID_INPUT
            - CURF_SQL_REJECTED
            - CURF_QUERY_TIMEOUT
            - CURF_TOO_MANY_QUERIES
            - CURF_UNAUTHENTICATED
            - CURF_FORBIDDEN
            - CURF_SEGREGATION_OF_DUTIES
            - CURF_NOT_FOUND
            - CURF_STALE_VERSION
            - CURF_CONFLICT
            - CURF_ROW_LIMIT
            - CURF_EXPORT_UNAVAILABLE
            - CURF_CONNECTION_FAILED
            - CURF_INTERNAL
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ReportDefinition:
      type: object
      required: [version, name, pages]
      description: |
        Curf's report definition, stored as sent (pages, blocks, themes and the rest are presentation the engine does not
        read). The engine reads `parameters` and `dataSources`, and which queries blocks use (any `queryId` or `*QueryId`
        value; `config.drilldown.queryId` marks drill-only queries). Each `dataSources[]` entry needs `id`, `name`,
        `dataSourceId` and exactly one of:

        - `engine`: `{viewId, columns?, filters?, groupBy?, aggregates?, orderBy?, limit?, params?}`, a published view plus
          structured parts, run as the caller. A value may be `{"$param": "name"}` to use a report parameter; a filter with
          `skipIfEmpty` is left out when its parameter is empty (Curf's "blank means no filter").
        - `sql`: a raw SELECT with `:name` parameters on the engine connection `dataSourceId`. Only for `view:manage` holders
          on connections that allow raw SQL; for everyone else the query is marked `accessDeniedNote`.

        Not supported yet: REST fields, `joins` and `attaches`.
      properties:
        version: { type: integer, enum: [1] }
        name: { type: string, minLength: 1, maxLength: 200 }
        parameters:
          type: array
          items:
            type: object
            required: [name, type]
            properties:
              name: { type: string, pattern: "^[A-Za-z_][A-Za-z0-9_]{0,63}$", description: Must not start with `__`. }
              label: { type: string }
              type: { type: string, enum: [string, number, date, dateRange, boolean, select] }
              required: { type: boolean }
              default: { type: [string, number, boolean] }
              options:
                type: array
                items:
                  type: object
                  properties:
                    value: { type: string }
                    label: { type: string }
        dataSources: { type: array, maxItems: 100, items: { type: object } }
        pages: { type: array, minItems: 1, items: { type: object } }
      additionalProperties: true
    ReportRequest:
      type: object
      required: [definition]
      properties:
        definition: { $ref: "#/components/schemas/ReportDefinition" }
        runRoles: { type: array, items: { type: string }, description: Roles that may run the report. Editors always may. }
        version: { type: integer, description: Required on update. }
        note: { type: string, maxLength: 500, description: Describes the change in the history. }
    Report:
      type: object
      required: [id, name, definition, runRoles, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        definition: { $ref: "#/components/schemas/ReportDefinition" }
        runRoles: { type: array, items: { type: string } }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
...
    RunnableReport:
      type: object
      required: [id, name, version]
      description: A report as someone who may only run it sees it. Queries are reduced to `id` and `name` when a definition is included.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        version: { type: integer }
        definition: { $ref: "#/components/schemas/ReportDefinition" }
...
    Provenance:
      type: object
      required: [queryId, queryName, queryHash, runAt, durationMs, rowCount, dataHash, dataSourceName, dataSourceKind]
      description: |
        Where one query's data came from, in the shape Curf records. `queryHash` and `dataHash` are computed by Curf's own
        algorithms (SHA-256, 24 hex characters, `sha256:` prefix), so they equal Curf's for the same SQL, parameters and
        rows. Date columns are plain dates here, so a result with date columns hashes differently from Curf's.
      properties:
        queryId: { type: string }
        queryName: { type: string }
        queryHash: { type: string, pattern: "^sha256:[0-9a-f]{24}$" }
        runAt: { type: string, format: date-time }
        durationMs: { type: integer, format: int64 }
        rowCount: { type: integer }
        dataHash: { type: string, pattern: "^sha256:[0-9a-f]{24}$" }
        dataSourceName: { type: string }
        dataSourceKind: { type: string, description: "`postgres` or `mysql` (MariaDB is `mysql`), as Curf names them." }
        accessDeniedNote: { type: string, description: The caller may not use this data source; the dataset is empty. }
        executionError: { type: string, maxLength: 500, description: The query failed; the dataset is empty and the rest of the report ran. }
        truncated: { type: boolean, description: Present and true when the row or byte cap cut the result short. }
    RunResult:
      type: object
      required: [dataset, provenance, params, reportVersion, asOf]
      properties:
        dataset:
          type: object
          description: Rows keyed by column name, one entry per query that ran. Numerics and (on PostgreSQL) bigints are text, as Curf's drivers return them.
          additionalProperties:
            type: array
            items: { type: object, additionalProperties: true }
        provenance:
          type: object
          additionalProperties: { $ref: "#/components/schemas/Provenance" }
        params: { type: object, description: Every declared parameter as applied, with its default and type conversion. }
        reportVersion: { type: integer }
        asOf: { type: string, format: date-time }
...